reformat_slice: Handle single integers too.
Adds support for handling single integers as well. Does this by
converting them to single element slices and thus handling them as
regular slices. At the very end, they are converted back to an integer
so they can still be used to index something per usual. Unlike slices
some indexes are not permissible. So if an integer index is converted to
an empty slice, this is raised as an index error.