reformat_slice: Drop unreachable/unused condition.
We encounter this condition earlier in the code. However, when we
encounter it we make change to the state such that the condition no
longer exists. So there is no reason to check for it later. Further we
weren't using this condition at all. As a result, it is completely
reasonable to drop it.