reformat_slice: Fixing out-of-bounds handling.
There were a couple of out-of-bounds cases with start and stop where
they did not effect each other. However, they were part of an `if/elif`
branch statement anyways. Even though both statements could be true and
executed separately, executing one was blocking the other. This patch
fixes it so that they both can be true and acted upon in spite of the
other case.