reformat_slice: Avoid comparisons with None.
In some cases the stop value needs to be `None`. This is needed as in
these cases there is no integer that stop can be that meets the
requirements. However, this means we need to guard about having `None`
show up in comparisons. This commit adds a `None` check for the stop
value and verifies that the stop value is not `None` in any case where
it is known to be possible.