reformat_slice: Handle an empty slice caused by stop.
Just as a stop value of 0 for a slice with a positive step size will
result in an empty slice, a stop value of -1 or positive equivalent
(given the length is known) will result in an empty slice given a
negative step size. This change handles that case and converts the slice
to a trivial empty slice to make it easier to recognize.