Add `split_multindex`
Adds a utility function, `split_multindex`, for splitting up slices with
multiple indices into a list of slices each with a single index. This is
handy when working with libraries that accept slices that have a single
index, but balk when multiple indices are provided. In these cases, one
can split up the multiple indices into multiple slices, which can then
be iterated over. Construction of the total result is up to user.