Add `index_to_slice`
Provides a simple formatting function that allows an integer index to be
converted to a slice. It is worth noting that a single integer index
does not behave the same as slice containing one index. The former
results in reduction of a dimension; whereas, the latter keeps the
dimension, but reduces it to a singleton dimension.