Restrict to one integral sequence per slice
There are a lot of subtle cases when support multiple integral
sequences, which we are not currently supporting correctly. So as to
avoid potentially providing an incorrect result, simply raise an
exception if multiple integral sequences show up in a slices and note
that only one is supported in a slice. Hopefully we can fix this and
relaxing this constraint in the future.