split_blocks: Add optional block index.
In some cases, it is handy to have a positional index for each block to
know where blocks are relative to each other. Having these indices can
be helpful when stitching the results back in some cases. This change
adds a flag so that we can optionally get this index as the first
argument. Also add a warning to indicate that this flag's default
behavior is going to change to `True` in the next minor release.