Require positive definite shape for `split_blocks`.
Make an assertion about the space shape. Namely require it to contain
only positive values. There should be no 0 values or the shape is
trivial and thus doesn't need to be divided into blocks. Further
negative values have no meaning in this context as they are not relative
to anything.