Assert -1/positive for block shape in split_blocks
Make an assertion about the block shape. Namely all values must be
positive or -1. Zero values are unacceptable as they would yield an
infinite number of trivial blocks. -1 is only allowed to extend to the
full length of the dimension. Potentially other negative values could be
made to work, but seem to be of dubious value for a block shape.